+++ src/collective/elastic/plone/restapi/schema.py
@@ -1,16 +1,22 @@
 from .service import Service
+from plone import api
 from plone.behavior.interfaces import IBehavior
 from Products.CMFCore.interfaces import ITypeInformation
 from zope.component import getAllUtilitiesRegisteredFor
 from zope.schema import getFieldsInOrder
 
 
+try:
+    from plone.app.multilingual.interfaces import ILanguageIndependentField
+except ImportError:
+    ILanguageIndependentField = None
+
+
 def dottedname(obj):
     module = obj.__class__.__module__
     if module is None or module == str.__class__.__module__:
         return obj.__class__.__name__  # Avoid reporting __builtin__
-    else:
-        return module + "." + obj.__class__.__name__
+    return module + "." + obj.__class__.__name__
 
 
 def _field(field, name=None):
@@ -21,6 +27,15 @@ def _field(field, name=None):
     value_type = getattr(field, "value_type", None)
     if value_type:
         record["value_type"] = _field(value_type)
+    if ILanguageIndependentField is not None and ILanguageIndependentField.providedBy(
+        field
+    ):
+        # p.a.multilingual available and language independent field
+        record["languages"] = []
+    else:
+        record["languages"] = api.portal.get_registry_record(
+            "plone.available_languages"
+        )
     return record